+++ MilkGlass/imgui.ini
@@ -1,6 +1,6 @@
 [Window][Editor Dockspace]
 Pos=0,0
-Size=2498,1417
+Size=1280,720
 Collapsed=0
 
 [Window][Debug##Default]
@@ -10,25 +10,25 @@ Collapsed=0
 
 [Window][ Scene Hierarchy]
 Pos=0,24
-Size=380,835
+Size=380,416
 Collapsed=0
 DockId=0x00000005,0
 
 [Window][ Properties]
-Pos=0,861
-Size=380,556
+Pos=0,442
+Size=380,278
 Collapsed=0
 DockId=0x00000006,0
 
 [Window][ Viewport]
 Pos=382,24
-Size=1734,1393
+Size=516,696
 Collapsed=0
 DockId=0x00000003,0
 
 [Window][ Stats]
-Pos=2118,24
-Size=380,1393
+Pos=900,24
+Size=380,696
 Collapsed=0
 DockId=0x00000004,0
 
@@ -48,7 +48,7 @@ Size=533,170
 Collapsed=0
 
 [Docking][Data]
-DockSpace     ID=0x7FE8E11A Window=0x26454203 Pos=0,47 Size=2498,1393 Split=X
+DockSpace     ID=0x7FE8E11A Window=0x26454203 Pos=675,443 Size=1280,696 Split=X
   DockNode    ID=0x00000001 Parent=0x7FE8E11A SizeRef=380,1417 Split=Y
     DockNode  ID=0x00000005 Parent=0x00000001 SizeRef=748,849 Selected=0x863995C4
     DockNode  ID=0x00000006 Parent=0x00000001 SizeRef=748,566 Selected=0xB41839C4
