+++ Milky/src/Milky/Scene/Components/Components.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "ComponentsBasic.h"
+#include "ComponentsVisual.h"
+#include "ComponentsLogic.h"
