+++ Milky/src/Milky/Scene/Components/ComponentsLogic.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "Milky/Scene/ScriptableEntity.h"
+
+namespace Milky {
+
+	struct NativeScriptComponent
+	{
+		ScriptableEntity* Instance = nullptr;
+
+		ScriptableEntity* (*InstantiateScript)();
+		void (*DestroyScript)(NativeScriptComponent*);
+
+		template<typename T>
+		void Bind()
+		{
+			InstantiateScript = []() { return static_cast<ScriptableEntity*>(new T()); };
+			DestroyScript = [](NativeScriptComponent* nsc) { delete nsc->Instance; nsc->Instance = nullptr; };
+		}
+	};
+
+}
